refactor(core/xref): run as module
