+++ profiles/w3c-common.js
@@ -51,6 +51,7 @@ const modules = [
   import("../src/core/webidl.js"),
   import("../src/core/biblio.js"),
   import("../src/core/link-to-dfn.js"),
+  import("../src/core/xref.js"),
   import("../src/core/data-cite.js"),
   import("../src/core/webidl-index.js"),
   import("../src/core/render-biblio.js"),
